udpated routes and api controllers
